5: Suche über Eingabemaske (http://github.com/fb71/polymap4-alkis/issues/issue/5)   - FlurstueckFilterPage: eingebaut, funktioniert grundsätzlich   - FIX: FlurstueckTransformer hat fälschlich ein " " vor alle felder gebaut, weshalb die suche über felder nicht funktioniert   - new form API eingebaut

+++ plugins/org.polymap.alkis/src/org/polymap/alkis/model/AX_Flurstueck_Kerndaten.java
@@ -105,7 +105,7 @@ public abstract class AX_Flurstueck_Kerndaten
      * zwei Nachkommastellen geführt werden, ansonsten ohne Nachkommastellen.
      */
     @NameInStore("amtlicheflaeche")
-    public Property<Float>                     amtlicheFlaeche;
+    public Property<Double>                     amtlicheFlaeche;
 
     /**
      * 'Flurnummer' ist die von der Katasterbehörde zur eindeutigen Bezeichnung
